bug: fix multiple traits not being gained for the same level
There is a bug where for npcs, if you put multiple traits growth for the
same level, only the last one will be awarded upon a level up.

Fix this by making the traits a Map<Integer, List<Trait>> to support
multiple traits at the same level.

Also refactor the levelup system to be in the Growth class.