Changing a character's initial gender now removes nonconforming genitals, while still respecting that the "genitals" field overrides everything. Example: A character has a built-in gender of male, with a massive cybernetic cock, but is configured to be female. The character now has no cock and a basic pussy.
Changing a character's initial gender now also keeps non-default characteristics of that character's built-in genital settings, as long as they still fit into the new gender. Example: A character has a built-in gender of male, with a massive cybernetic cock, but is configured to be herm. The character now has a massive cybernetic cock and a basic pussy.

Configured archetypes are applied after the genitals have been modified.